% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Concatenate a vector of strings}
\usage{
collapse(vector, collapse = " ")
}
\arguments{
\item{vector}{Character vector.}

\item{collapse}{A separator.}
}
\value{
character
}
\description{
Helper function for pretty printing vectors. Avoids repetitive paste0 code. If only one element does not add the separator at the end.
}
\examples{
collapse(letters, ", ")
collapse(letters[1], ", ")
}
