% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger}
\alias{Logger}
\title{Logger}
\description{
An R6 class for flexible logging with customizable output.
}
\examples{
# Create a basic logger
logger <- Logger$new()
logger$info("This is an info message")
logger$warn("This is a warning")
logger$error("This is an error")

# Create a logger with custom settings
custom_logger <- Logger$new(
  level = LogLevel$WARNING,
  save_to_file = TRUE,
  file_path = tempfile("log_"),
  print_fn = message
)
custom_logger$info("This won't be logged")
custom_logger$warn("This will be logged")

# Change log level
custom_logger$set_level(LogLevel$INFO)
custom_logger$info("Now this will be logged")

## ------------------------------------------------
## Method `Logger$new`
## ------------------------------------------------

logger <- Logger$new(level = LogLevel$WARNING, save_to_file = TRUE, file_path = "log.txt")

## ------------------------------------------------
## Method `Logger$set_level`
## ------------------------------------------------

logger <- Logger$new()
logger$set_level(LogLevel$WARNING)

## ------------------------------------------------
## Method `Logger$error`
## ------------------------------------------------

logger <- Logger$new()
logger$error("An error occurred", data = list(x = 1), error = simpleError("Oops!"))

## ------------------------------------------------
## Method `Logger$warn`
## ------------------------------------------------

logger <- Logger$new()
logger$warn("This is a warning", data = list(reason = "example"))

## ------------------------------------------------
## Method `Logger$info`
## ------------------------------------------------

logger <- Logger$new()
logger$info("Operation completed successfully", data = list(duration = 5.2))
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Logger-new}{\code{Logger$new()}}
\item \href{#method-Logger-set_level}{\code{Logger$set_level()}}
\item \href{#method-Logger-error}{\code{Logger$error()}}
\item \href{#method-Logger-warn}{\code{Logger$warn()}}
\item \href{#method-Logger-info}{\code{Logger$info()}}
\item \href{#method-Logger-clone}{\code{Logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-new"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-new}{}}}
\subsection{Method \code{new()}}{
Create a new Logger object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(
  level = LogLevel$INFO,
  save_to_file = FALSE,
  file_path = "",
  print_fn = function(x) cat(x, "\\n")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The minimum log level to output. Default is LogLevel$INFO.}

\item{\code{save_to_file}}{Logical; whether to save logs to a file. Default is FALSE.}

\item{\code{file_path}}{Character; path to the log file if save_to_file is TRUE.}

\item{\code{print_fn}}{Function; custom print function to use for console output.
Should accept a single character string as input. Default uses cat with a newline.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Logger` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new(level = LogLevel$WARNING, save_to_file = TRUE, file_path = "log.txt")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-set_level"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-set_level}{}}}
\subsection{Method \code{set_level()}}{
Set the minimum log level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$set_level(level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The new minimum log level to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$set_level(LogLevel$WARNING)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-error"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-error}{}}}
\subsection{Method \code{error()}}{
Log an error message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$error(msg, data = NULL, error = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Character; the error message to log.}

\item{\code{data}}{Optional; additional data to include in the log entry.}

\item{\code{error}}{Optional; an error object to include in the log entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$error("An error occurred", data = list(x = 1), error = simpleError("Oops!"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-warn"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-warn}{}}}
\subsection{Method \code{warn()}}{
Log a warning message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$warn(msg, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Character; the warning message to log.}

\item{\code{data}}{Optional; additional data to include in the log entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$warn("This is a warning", data = list(reason = "example"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-info"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-info}{}}}
\subsection{Method \code{info()}}{
Log an info message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$info(msg, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Character; the info message to log.}

\item{\code{data}}{Optional; additional data to include in the log entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$info("Operation completed successfully", data = list(duration = 5.2))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
