% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stenographer.R
\name{valid_log_level}
\alias{valid_log_level}
\title{Check if an object is a valid log level}
\usage{
valid_log_level(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if an object is a valid log level
}
