% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger}
\alias{Logger}
\title{Logger}
\description{
An R6 class for flexible logging with customisable output, message formatting, and context.
}
\examples{
# Create a basic logger
logger <- Logger$new()
logger$info("This is an info message")
logger$warn("This is a warning")
logger$error("This is an error")

# Create a logger with custom settings, message formatting, and context
custom_logger <- Logger$new(
  level = LogLevel$WARNING,
  file_path = tempfile("log_"),
  print_fn = function(x) message(paste0("Custom: ", x)),
  format_fn = function(level, msg) paste0("Hello prefix: ", msg),
  context = list(program = "MyApp")
)
custom_logger$info("This won't be logged")
custom_logger$warn("This will be logged with a custom prefix")

# Change log level and update context
custom_logger$set_level(LogLevel$INFO)
custom_logger$update_context(list(user = "John"))
custom_logger$info("Now this will be logged with a custom prefix and context")

## ------------------------------------------------
## Method `Logger$set_level`
## ------------------------------------------------

logger <- Logger$new()
logger$set_level(LogLevel$WARNING)

## ------------------------------------------------
## Method `Logger$error`
## ------------------------------------------------

logger <- Logger$new()
logger$error("An error occurred", data = list(x = 1), error = "Oops!")

## ------------------------------------------------
## Method `Logger$warn`
## ------------------------------------------------

logger <- Logger$new()
logger$warn("This is a warning", data = list(reason = "example"))

## ------------------------------------------------
## Method `Logger$info`
## ------------------------------------------------

logger <- Logger$new()
logger$info("Operation completed successfully", data = list(duration = 5.2))
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Logger-new}{\code{Logger$new()}}
\item \href{#method-Logger-set_level}{\code{Logger$set_level()}}
\item \href{#method-Logger-update_context}{\code{Logger$update_context()}}
\item \href{#method-Logger-clear_context}{\code{Logger$clear_context()}}
\item \href{#method-Logger-get_context}{\code{Logger$get_context()}}
\item \href{#method-Logger-error}{\code{Logger$error()}}
\item \href{#method-Logger-warn}{\code{Logger$warn()}}
\item \href{#method-Logger-info}{\code{Logger$info()}}
\item \href{#method-Logger-clone}{\code{Logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-new"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-new}{}}}
\subsection{Method \code{new()}}{
Create a new Logger object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(
  level = LogLevel$INFO,
  file_path = NULL,
  db_conn = NULL,
  table_name = "LOGS",
  print_fn = function(x) cat(x, "\\n"),
  format_fn = function(level, msg) msg,
  context = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The minimum log level to output. Default is LogLevel$INFO.}

\item{\code{file_path}}{Character; the path to a file to save log entries to. Default is NULL.}

\item{\code{db_conn}}{DBI connection object; an existing database connection. Default is NULL.}

\item{\code{table_name}}{Character; the name of the table to log to in the database. Default is "LOGS".}

\item{\code{print_fn}}{Function; custom print function to use for console output.
Should accept a single character string as input. Default uses cat with a newline.}

\item{\code{format_fn}}{Function; custom format function to modify the log message.
Should accept level and msg as inputs and return a formatted string.}

\item{\code{context}}{List; initial context for the logger. Default is an empty list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Logger` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-set_level"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-set_level}{}}}
\subsection{Method \code{set_level()}}{
Set the minimum log level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$set_level(level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The new minimum log level to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$set_level(LogLevel$WARNING)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-update_context"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-update_context}{}}}
\subsection{Method \code{update_context()}}{
Update the logger's context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$update_context(new_context)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_context}}{A list of new context items to add or update}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-clear_context"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-clear_context}{}}}
\subsection{Method \code{clear_context()}}{
Clear the logger's context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clear_context()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-get_context"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-get_context}{}}}
\subsection{Method \code{get_context()}}{
Get the current context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$get_context()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-error"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-error}{}}}
\subsection{Method \code{error()}}{
Log an error message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$error(msg, data = NULL, error = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Character; the error message to log.}

\item{\code{data}}{Optional; additional data to include in the log entry.}

\item{\code{error}}{Optional; an error object to include in the log entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$error("An error occurred", data = list(x = 1), error = "Oops!")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-warn"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-warn}{}}}
\subsection{Method \code{warn()}}{
Log a warning message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$warn(msg, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Character; the warning message to log.}

\item{\code{data}}{Optional; additional data to include in the log entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$warn("This is a warning", data = list(reason = "example"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-info"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-info}{}}}
\subsection{Method \code{info()}}{
Log an info message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$info(msg, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Character; the info message to log.}

\item{\code{data}}{Optional; additional data to include in the log entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{logger <- Logger$new()
logger$info("Operation completed successfully", data = list(duration = 5.2))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
