% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\docType{data}
\name{LogLevel}
\alias{LogLevel}
\title{Log Levels}
\format{
An object of class \code{list} of length 3.
}
\usage{
LogLevel
}
\description{
Defines the available log levels for the Logger class.
}
\keyword{datasets}
