% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stenographer.R
\docType{data}
\name{LogLevel}
\alias{LogLevel}
\title{Logging Level}
\format{
An object of class \code{list} of length 4.
}
\usage{
LogLevel
}
\description{
Defines standard logging levels for controlling message output granularity.
Use as a configuration for the `Stenographer` class to control which messages
are logged.

A list with four integer elements:
\describe{
  \item{OFF}{(-1) Disables all logging}
  \item{ERROR}{(0) Logs only errors}
  \item{WARNING}{(1) Logs errors and warnings}
  \item{INFO}{(2) Logs all messages}
}
}
\examples{
# Check logging levels
LogLevel$OFF     # -1
LogLevel$ERROR   # 0
LogLevel$WARNING # 1
LogLevel$INFO    # 2

}
\keyword{datasets}
